brought changes from https://github.com/OneSignal/OneSignal-Android-SDK/pull/2539/changes

+++ OneSignalSDK/onesignal/otel/src/main/java/com/onesignal/otel/IOtelPlatformProvider.kt
@@ -30,7 +30,7 @@ interface IOtelPlatformProvider {
     val onesignalId: String?
     val pushSubscriptionId: String?
     val appState: String // "foreground" or "background"
-    val processUptime: Double // in seconds
+    val processUptime: Long // in milliseconds
     val currentThreadName: String
 
     // Crash-specific configuration
